+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This section is for maintaining a changelog for all breaking changes for the cli
 ## [Unreleased 2.x]
 
 ### Added
+- Added support for indexing and search index settings ([#667](https://github.com/opensearch-project/opensearch-java/pull/667))
 
 ### Dependencies